src/requirements.txt
fastapi
pydantic
uvicorn
Pillow
python-multipart
pytest
httpx